refactor: Refactor the project structure and optimize the code
